Changed amount from Float to Int due to type inaccuracy

+++ app/src/main/java/com/example/moneymanager/data/model/Account.kt
@@ -9,6 +9,6 @@ data class Account(
     val id: Long,
     val group: String,
     val name: String,
-    val amount: Float,
+    val amount: Int,
     val includeInTotals: Boolean
 )
